requirements.txt
django==4.2.25
djangorestframework==3.16.1
pytest==8.4.2
pytest-django==4.11.1
pytest-mock==3.15.1
freezegun==1.5.5
PyJWT==2.10.1